File rename in vehicles
