+++ vehicles/src/common/enum/user-role.enum.ts
