+++ vehicles/src/modules/company-user-management/users/dto/response/read-user.dto.ts
@@ -4,7 +4,7 @@ import {
   ClientUserRole,
   IDIRUserRole,
   UserRole,
-} from '../../../../../common/enum/user-auth-group.enum';
+} from '../../../../../common/enum/user-role.enum';
 import { UserStatus } from '../../../../../common/enum/user-status.enum';
 import { ReadContactDto } from '../../../../common/dto/response/read-contact.dto';
