+++ vehicles/src/modules/company-user-management/users/entities/user.entity.ts
@@ -14,7 +14,7 @@ import { Directory } from '../../../../common/enum/directory.enum';
 import {
   GenericUserRole,
   IDIRUserRole,
-} from '../../../../common/enum/user-auth-group.enum';
+} from '../../../../common/enum/user-role.enum';
 import { UserStatus } from '../../../../common/enum/user-status.enum';
 
 @Entity({ name: 'ORBC_USER' })
