+++ vehicles/src/modules/permit-application-payment/application/profile/application.profile.ts
@@ -16,7 +16,7 @@ import { Directory } from '../../../../common/enum/directory.enum';
 import {
   UserRole,
   IDIR_USER_ROLE_LIST,
-} from '../../../../common/enum/user-auth-group.enum';
+} from '../../../../common/enum/user-role.enum';
 import { doesUserHaveAuthGroup } from '../../../../common/helper/auth.helper';
 import { Permit } from '../../permit/entities/permit.entity';
