+++ vehicles/src/modules/shopping-cart/profile/shopping-cart.profile.ts
@@ -8,7 +8,7 @@ import {
 import { AutomapperProfile, InjectMapper } from '@automapper/nestjs';
 import { Injectable } from '@nestjs/common';
 
-import { UserRole } from '../../../common/enum/user-auth-group.enum';
+import { UserRole } from '../../../common/enum/user-role.enum';
 import { getApplicantDisplay } from '../../../common/helper/permit-application.helper';
 import { Permit as Application } from '../../permit-application-payment/permit/entities/permit.entity';
 import { ReadShoppingCartDto } from '../dto/response/read-shopping-cart.dto';
